rag_glaucoma_screening/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.36.0
Pillow>=10.1.0
numpy>=1.24.3
pandas>=2.1.3
scikit-learn>=1.3.2
matplotlib>=3.8.2
seaborn>=0.13.0
tqdm>=4.66.1
faiss-cpu>=1.7.4
kaggle>=1.5.16
huggingface-hub>=0.20.0
